+++ src/activitypub/handlers/update.rs
@@ -142,6 +142,7 @@ async fn handle_update_person(
         &MediaStorage::from(config),
         profile,
         activity.object,
+        false,
     ).await?;
     Ok(Some(PERSON))
 }